Add "dir" member to ConnectionRef, an enum that can have 3 values: kUnspecified (default), kInput and kOutput. Expect kUnspecified to only be used for testing, while confgen will restrict generated ConnectionRef instances to kInput or kOutput. Check directionality in get_sender and get_receiver.
Make Sender and Receiver inherit from NamedObject so that queue-like
usages can continue to call get_name().

+++ unittest/performance_test.cxx
@@ -60,8 +60,8 @@ struct ConfigurationTestFixture
   dunedaq::iomanager::IOManager iom;
   dunedaq::iomanager::ConnectionRef conn_ref;
   dunedaq::iomanager::ConnectionRef queue_ref;
-  const int n_sends = 10000;
-  const int message_size = 55680;
+  const size_t n_sends = 10000;
+  const size_t message_size = 55680;
 };
 
 BOOST_FIXTURE_TEST_CASE(CallbackRegistrationNetwork, ConfigurationTestFixture)
